Fix leak on move_to_next_command_list in main.c
