+++ src/utils.c
@@ -6,25 +6,11 @@
 /*   By: jongpark <jongpark@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/09 14:51:36 by dason             #+#    #+#             */
-/*   Updated: 2021/11/22 11:10:55 by jongpark         ###   ########.fr       */
+/*   Updated: 2021/11/29 17:27:47 by dason            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-void	free_double_pointer(char ***s)
-{
-	int		i;
-
-	i = -1;
-	while ((*s)[++i])
-	{
-		free((*s)[i]);
-		(*s)[i] = NULL;
-	}
-	free(*s);
-	*s = NULL;
-}
 
 bool	is_blank(char *str)
 {
